Check if an upstream demuxer provided a default kid.
Smooth streaming demuxers do not send boxes containing metadata about
the stream. They have to send metadata "out-of-band".

For piff encoded streams, if no box has been found containing a default
kid for the sample, check if an upstream demuxer has provided one via a
protection event, and use this as a fallback.